refactor: change function in task 04 to recursive

+++ 04-copy-directory/index.js
@@ -5,14 +5,14 @@ const path = require('path');
 const pathFrom = path.join(__dirname, 'files');
 const pathTo = path.join(__dirname, 'files-copy');
 
-fs.mkdir(pathTo, { recursive: true }, error => {
-  if(error) {
-    console.log(`There occurred an error: ${error.message}`);
-  }
-});
 copyFilesFromDirectory(pathFrom, pathTo); 
 
 async function copyFilesFromDirectory(pathFrom, pathTo) {
+  fs.mkdir(pathTo, { recursive: true }, error => {
+    if(error) {
+      console.log(`There occurred an error: ${error.message}`);
+    }
+  });
   const files = await fs.readdir(pathFrom, { withFileTypes: true }, error => {
     if(error) {
       console.log(`There occurred an error: ${error.message}`); 
@@ -20,14 +20,19 @@ async function copyFilesFromDirectory(pathFrom, pathTo) {
   });
   
   for(let file of files) {
-    copyFile(path.join(pathFrom, file.name), path.join(pathTo, file.name), error => {
-      if(error) {
-        console.log(`There occurred an error: ${error.message}`); 
-      }
-    });
+    if(file.isFile()) {
+      copyFile(path.join(pathFrom, file.name), path.join(pathTo, file.name), error => {
+        if(error) {
+          console.log(`There occurred an error: ${error.message}`); 
+        }
+      });
+    }
+    else if(file.isDirectory()) {
+      copyFilesFromDirectory(path.resolve(pathFrom, file.name), path.resolve(pathTo, file.name));
+    }  
   }
   setTimeout(() => {
-    console.log(`All the files have been successfully copied to the 'file-copy' directory.`);
+    console.log(`All the files have been successfully copied.`);
   }, 100);
 }
